+++ config/config.exs
@@ -1,0 +1,4 @@
+import Config
+
+config :ex_cldr,
+  default_backend: MyApp.Cldr